Relocalizada classe de teste de controlador para package de controller.
